Modify index's base href-path and introduced deployUrl in angular-cli.json
In previous versions I mistakingly changed the base Tag in the Index.html so it would point to the path with the bundle scripts. The problem is that the angular router needs base set to \ for correct navigation.

The alternative is to add a 'deployUrl:' value to the apps section inside the .angular-cli.json, which then adjusts the paths for the scripts on new build.